requirements.txt
streamlit
langchain
langchain-google-genai
langchain-core
langgraph
python-dotenv
requests
pydantic